add title to hero
