+++ src/components/Hero.jsx
@@ -1,40 +1,29 @@
 import React from "react";
-import { toast } from "react-toastify";
 
 export default function Hero(props) {
-  const notify = () =>
-    toast.success("🦄 Wow so easy!", {
-      position: "bottom-left",
-      autoClose: 5000,
-      hideProgressBar: false,
-      closeOnClick: true,
-      pauseOnHover: true,
-      draggable: true,
-      progress: undefined,
-      theme: "light",
-    });
   return (
     <div className="relative bg-[url(./Assets/img/Web/heroImage.webp)] bg-cover bg-center bg-no-repeat">
       <div className="absolute inset-0 bg-white/75 sm:bg-transparent sm:bg-gradient-to-r sm:from-white/95 sm:to-white/25"></div>
 
       <div className="relative mx-auto max-w-screen-xl px-4 py-32 sm:px-6 lg:flex lg:h-screen lg:items-center lg:px-8">
         <div className="max-w-xl text-center sm:text-left">
-          <span className="text-2xl font-bold sm:text-8xl py-5">
-            PESU <span className="text-blue-700">Cafe</span>
-            <button onClick={notify}>Notify !</button>
-          </span>
+          <div className="my-3">
+            <span className="text-2xl font-bold sm:text-8xl text-orange-500">
+              PESU
+            </span>
+            &nbsp;&nbsp;
+            <span className="text-blue-900 font-semibold text-8xl">Cafe</span>
+          </div>
           <h1 className="text-3xl font-extrabold sm:text-5xl">
             Ordering Snacks
             <strong className="block font-extrabold text-blue-700">
               made easier
             </strong>
           </h1>
-
           <p className="mt-4 max-w-lg sm:text-xl sm:leading-relaxed">
             Lorem ipsum dolor sit amet consectetur, adipisicing elit. Nesciunt
             illo tenetur fuga ducimus numquam ea!
           </p>
-
           <div className="mt-8 flex flex-wrap gap-4 text-center">
             <a
               href="#"
